Update comments for moved cart logic
