+++ mentions-legales.php
@@ -11,7 +11,7 @@ $page_title = "Mentions Légales";
 $page_description = "Consultez les mentions légales du site Verre & Image, informations sur l'éditeur, l'hébergement et la politique de confidentialité.";
 
 // Inclure l'en-tête commun à toutes les pages
-require_once 'templates/header.php';
+require_once 'src/templates/header.php';
 ?>
 
 <style>
@@ -93,5 +93,5 @@ require_once 'templates/header.php';
 
 <?php
 // Inclure le pied de page commun à toutes les pages
-require_once 'templates/footer.php';
+require_once 'src/templates/footer.php';
 ?>
