+++ mon_compte.php
@@ -8,7 +8,7 @@
 
 // Inclure l'initialisation pour accéder aux sessions et fonctions
 require_once 'config/init.php';
-require_once 'includes/fonctions.php';
+require_once 'src/functions.php';
 
 // Vérifier si l'utilisateur est connecté. Si non, le rediriger vers la page de connexion.
 if (!is_user_logged_in()) {
@@ -40,7 +40,7 @@ $page_title = "Mon Compte";
 $page_description = "Gérez vos informations personnelles et consultez l'historique de vos commandes sur Verre & Image.";
 
 // Inclure l'en-tête
-require_once 'templates/header.php';
+require_once 'src/templates/header.php';
 ?>
 
 <style>
@@ -162,5 +162,5 @@ require_once 'templates/header.php';
 
 <?php
 // Inclure le pied de page
-require_once 'templates/footer.php';
+require_once 'src/templates/footer.php';
 ?>
